refactor; move questionry mocking
